Relocate ddl impl
